+++ DatadogSessionReplay/Sources/Recorder/ViewTreeSnapshotProducer/ViewTreeSnapshot/NodeRecorders/UITextViewRecorder.swift
@@ -72,10 +72,10 @@ internal struct UITextViewWireframesBuilder: NodeWireframesBuilder {
         let bottom = max(contentRect.height - attributes.frame.height - top, 0)
         let right = max(contentRect.width - attributes.frame.width - left, 0)
         return SRContentClip(
-            bottom: Int64(withNoOverflow: bottom),
-            left: Int64(withNoOverflow: left),
-            right: Int64(withNoOverflow: right),
-            top: Int64(withNoOverflow: top)
+            bottom: Int(withNoOverflow: bottom),
+            left: Int(withNoOverflow: left),
+            right: Int(withNoOverflow: right),
+            top: Int(withNoOverflow: top)
         )
     }
